+++ test/integration/harness.js
@@ -1,0 +1,5 @@
+const endPoint = "http://localhost:3000";
+
+export {
+  endPoint
+};